bootloader_api::info

Struct BootInfo

#[non_exhaustive]

#[repr(C)]
pub struct BootInfo {
    pub api_version: ApiVersion,
    pub memory_regions: MemoryRegions,
    pub framebuffer: Optional<FrameBuffer>,
    pub physical_memory_offset: Optional<u64>,
    pub recursive_index: Optional<u16>,
    pub rsdp_addr: Optional<u64>,
    pub tls_template: Optional<TlsTemplate>,
    pub ramdisk_addr: Optional<u64>,
    pub ramdisk_len: u64,
    pub kernel_addr: u64,
    pub kernel_len: u64,
    pub kernel_image_offset: u64,
    /* private fields */
}

This structure represents the information that the bootloader passes to the kernel.

The information is passed as an argument to the entry point. The entry point function must have the following signature:

extern "C" fn(boot_info: &'static mut BootInfo) -> !;

Note that no type checking occurs for the entry point function, so be careful to use the correct argument types. To ensure that the entry point function has the correct signature, use the [entry_point] macro.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

api_version: ApiVersion

The version of the bootloader_api crate. Must match the bootloader version.

memory_regions: MemoryRegions

A map of the physical memory regions of the underlying machine.

The bootloader queries this information from the BIOS/UEFI firmware and translates this information to Rust types. It also marks any memory regions that the bootloader uses in the memory map before passing it to the kernel. Regions marked as usable can be freely used by the kernel.

framebuffer: Optional<FrameBuffer>

Information about the framebuffer for screen output if available.

physical_memory_offset: Optional<u64>

The virtual address at which the mapping of the physical memory starts.

Physical addresses can be converted to virtual addresses by adding this offset to them.

The mapping of the physical memory allows to access arbitrary physical frames. Accessing frames that are also mapped at other virtual addresses can easily break memory safety and cause undefined behavior. Only frames reported as USABLE by the memory map in the BootInfo can be safely accessed.

Only available if the map-physical-memory config option is enabled.

recursive_index: Optional<u16>

The virtual address of the recursively mapped level 4 page table.

Only available if the map-page-table-recursively config option is enabled.

rsdp_addr: Optional<u64>

The address of the RSDP data structure, which can be use to find the ACPI tables.

This field is None if no RSDP was found (for BIOS) or reported (for UEFI).

tls_template: Optional<TlsTemplate>

The thread local storage (TLS) template of the kernel executable, if present.

ramdisk_addr: Optional<u64>

Ramdisk address, if loaded

ramdisk_len: u64

Ramdisk image size, set to 0 if addr is None

kernel_addr: u64

Physical address of the kernel ELF in memory.

kernel_len: u64

Size of the kernel ELF in memory.

kernel_image_offset: u64

Virtual address of the loaded kernel image.

Implementations

impl BootInfo

pub fn new(memory_regions: MemoryRegions) -> Self

Create a new boot info structure with the given memory map.

The other fields are initialized with default values.